Move Screen component into ui/
